Update GitHub Actions and SonarCloud Configuration
Description
This PR updates several GitHub Actions workflows and the SonarCloud configuration for the project.

Changes
Updated the .gitignore file to ignore various log files, editor directories, and build output directories.

Updated the User First Interaction GitHub Actions workflow. The workflow posts a message when a user opens their first issue or makes their first pull request in the repository.

Updated the Deploy static content to Github Pages GitHub Actions workflow. The workflow is triggered on push events to the main branch and deploys the static content to GitHub Pages.

Updated the Mind Your Language GitHub Actions workflow. The workflow is triggered when an issue or a comment is created or edited, and checks for profanity in the content.

Updated the sonar-project.properties file for SonarCloud analysis. The project key is set to TheRadixHub_radix.network and the organization is set to radixhubsonarcloud.